requirements.txt
streamlit==1.41.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-google-genai==2.0.8
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
google-generativeai==0.8.3
python-dotenv==1.0.1
chromadb==0.5.0
PyMuPDF==1.25.2
docx2txt==0.8
python-docx==1.1.2
beautifulsoup4==4.13.3
numpy<2.0.0,>=1.24.0
pandas>=2.2.0


